projects/social_media_ads_simulation-lkrvavica/requirements.txt
requests
duckdb
numpy
pandas
python-dotenv
marimo
pyyaml
ipython
scikit-learn
google-generativeai
streamlit
pydantic
pydantic-settings
pytest
pytest-cov
playwright
google-search-results
